requirements.txt
asgiref==3.8.1
babel==2.16.0
beautifulsoup4==4.12.3
black==24.10.0
bs4==0.0.2
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Django==5.1.5
django-allauth==65.3.1
django-crispy-forms==2.3
django-currencies==0.10.1
django-debug-toolbar==5.0.1
django-environ==0.12.0
django-money==3.5.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
flake8==7.1.1
idna==3.10
inflection==0.5.1
jsonfield==2.1.1
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
psycopg2-binary==2.9.10
py-moneyed==3.0
pycodestyle==2.12.1
pyflakes==3.2.0
PyJWT==2.10.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.3
setuptools==75.8.0
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.2.0
urllib3==2.3.0
